+++ src/inference/gibbs.jl
