+++ src/inference/smc.jl
@@ -41,13 +41,13 @@ function SMC(n_particles::Int, space...)
 end
 SMC(alg::SMC, new_gid::Int) = SMC(alg.n_particles, alg.resampler, alg.resampler_threshold, alg.space, new_gid)
 
-Sampler(alg::SMC) = begin
-  info = Dict{Symbol, Any}()
-  info[:logevidence] = []
-  Sampler(alg, info)
+function Sampler(alg::SMC)
+    info = Dict{Symbol, Any}()
+    info[:logevidence] = []
+    return Sampler(alg, info)
 end
 
-step(model, spl::Sampler{<:SMC}, vi::VarInfo) = begin
+function step(model, spl::Sampler{<:SMC}, vi::VarInfo)
     particles = ParticleContainer{Trace}(model)
     vi.num_produce = 0;  # Reset num_produce before new sweep\.
     set_retained_vns_del_by_spl!(vi, spl)
@@ -67,23 +67,23 @@ step(model, spl::Sampler{<:SMC}, vi::VarInfo) = begin
     indx = randcat(Ws)
     push!(spl.info[:logevidence], particles.logE)
 
-    particles[indx].vi
+    return particles[indx].vi
 end
 
 ## wrapper for smc: run the sampler, collect results.
 function sample(model::Model, alg::SMC)
-  spl = Sampler(alg);
+    spl = Sampler(alg)
 
-  particles = ParticleContainer{Trace}(model)
-  push!(particles, spl.alg.n_particles, spl, VarInfo())
+    particles = ParticleContainer{Trace}(model)
+    push!(particles, spl.alg.n_particles, spl, VarInfo())
 
-  while consume(particles) != Val{:done}
-    ess = effectiveSampleSize(particles)
-    if ess <= spl.alg.resampler_threshold * length(particles)
-      resample!(particles,spl.alg.resampler)
+    while consume(particles) != Val{:done}
+      ess = effectiveSampleSize(particles)
+      if ess <= spl.alg.resampler_threshold * length(particles)
+        resample!(particles,spl.alg.resampler)
+      end
     end
-  end
-  w, samples = getsample(particles)
-  res = Chain(w, samples)
-
+    w, samples = getsample(particles)
+    res = Chain(w, samples)
+    return res
 end
